Break out config utils into separate module #7 #8
